test: 합계금액을 계산하는 실패하는 테스트 작성

+++ src/main/java/lotto/domain/Lottos.java
@@ -10,9 +10,9 @@ public final class Lottos {
   private final List<Lotto> lottos;
 
   public Lottos(Money money) {
-    int size = money.divideBy(PRICE).toInteger();
+    long size = money.divideBy(PRICE).toInteger();
 
-    lottos = new ArrayList<>(size);
+    lottos = new ArrayList<>((int) size);
     for (int i = 0; i < size; i++) {
       lottos.add(new Lotto());
     }
